refactor: :art: restructure test folder
