+++ __test__/integration/routes/v1predicates.test.js
@@ -1,4 +1,4 @@
-const app = require("../../../../src/app");
+const app = require("../../../src/app");
 const request = require('supertest');
 
 describe("Test /v1/predicates endpoint", () => {
