+++ __test__/integration/routes/v1predicates_by_api.test.js
@@ -1,4 +1,4 @@
-const app = require("../../../../src/app");
+const app = require("../../../src/app");
 const request = require('supertest');
 
 describe("Test /v1/smartapi/{smartapi_id}/predicates endpoint", () => {
