+++ __test__/integration/routes/v1query.test.js
@@ -1,10 +1,10 @@
-const app = require("../../../../src/app");
+const app = require("../../../src/app");
 const request = require('supertest');
 const fs = require("fs");
 var path = require('path');
 
 describe("Testing /v1/query endpoints", () => {
-    const invalid_example_folder = path.resolve(__dirname, "../../../../examples/v1/invalid");
+    const invalid_example_folder = path.resolve(__dirname, "../../../examples/v1/invalid");
     test("Input query graph that doesn't pass Swagger Validation should return 400 error", async () => {
         const InvalidInputQueryGraph = {
             message1: 1
