+++ __test__/integration/routes/v1query_by_api.test.js
@@ -1,18 +1,18 @@
-const app = require("../../../../src/app");
+const app = require("../../../src/app");
 const request = require('supertest');
 const fs = require("fs");
 const axios = require("axios");
 var path = require('path');
 
-describe("Testing /v1/team/{team_name}/query endpoints", () => {
-    const invalid_example_folder = path.resolve(__dirname, "../../../../examples/v1/invalid");
-    const example_foler = path.resolve(__dirname, '../../../../examples/v1');
+describe("Testing /v1/smartapi/{smartapi_id}/query endpoints", () => {
+    const invalid_example_folder = path.resolve(__dirname, "../../../examples/v1/invalid");
+    const example_foler = path.resolve(__dirname, '../../../examples/v1');
     test("Input query graph that doesn't pass Swagger Validation should return 400 error", async () => {
         const InvalidInputQueryGraph = {
             message1: 1
         };
         await request(app)
-            .post("/v1/team/Text Mining Provider/query/")
+            .post("/v1/smartapi/5be0f321a829792e934545998b9c6afe/query/")
             .send(InvalidInputQueryGraph)
             .set('Accept', 'application/json')
             .expect(400)
@@ -25,7 +25,7 @@ describe("Testing /v1/team/{team_name}/query endpoints", () => {
     test("Input query graph missing nodes definition should return 400", async () => {
         const query_with_nodes_undefined = JSON.parse(fs.readFileSync(path.join(invalid_example_folder, "query_graph_with_nodes_not_specified.json")));
         await request(app)
-            .post("/v1/team/Text Mining Provider/query/")
+            .post("/v1/smartapi/5be0f321a829792e934545998b9c6afe/query/")
             .send(query_with_nodes_undefined)
             .set('Accept', 'application/json')
             .expect(400)
@@ -38,7 +38,7 @@ describe("Testing /v1/team/{team_name}/query endpoints", () => {
     test("Input query graph missing edges definition should return 400 error", async () => {
         const query_with_edges_undefined = JSON.parse(fs.readFileSync(path.join(invalid_example_folder, "query_graph_with_edges_not_specified.json")));
         await request(app)
-            .post("/v1/team/Text Mining Provider/query/")
+            .post("/v1/smartapi/5be0f321a829792e934545998b9c6afe/query/")
             .send(query_with_edges_undefined)
             .set('Accept', 'application/json')
             .expect(400)
@@ -51,7 +51,7 @@ describe("Testing /v1/team/{team_name}/query endpoints", () => {
     test("Input query graph with nodes and edges mismatch should return 400 error", async () => {
         const query_with_nodes_and_edges_not_match = JSON.parse(fs.readFileSync(path.join(invalid_example_folder, "query_graph_with_nodes_and_edges_not_match.json")));
         await request(app)
-            .post("/v1/team/Text Mining Provider/query/")
+            .post("/v1/smartapi/5be0f321a829792e934545998b9c6afe/query/")
             .send(query_with_nodes_and_edges_not_match)
             .set('Accept', 'application/json')
             .expect(400)
@@ -61,10 +61,10 @@ describe("Testing /v1/team/{team_name}/query endpoints", () => {
             })
     })
 
-    test("Query to Text Mining KPs should have id resolution turned off", async () => {
+    test("Query to Text Mining Targeted Association KP should have id resolution turned off", async () => {
         const query = JSON.parse(fs.readFileSync(path.join(example_foler, "textmining/query_chemicals_related_to_gene_or_gene_product.json")));
         await request(app)
-            .post("/v1/team/Text Mining Provider/query/")
+            .post("/v1/smartapi/978fe380a147a8641caf72320862697b/query/")
             .send(query)
             .set('Accept', 'application/json')
             .expect(200)
@@ -75,19 +75,33 @@ describe("Testing /v1/team/{team_name}/query endpoints", () => {
             })
     })
 
-    test("Query to Service Provider KPs should have id resolution turned on", async () => {
+    test("Query to Text Mining Targeted Association KP should have id resolution turned off", async () => {
         const query = JSON.parse(fs.readFileSync(path.join(example_foler, "textmining/query_chemicals_related_to_gene_or_gene_product.json")));
         await request(app)
-            .post("/v1/team/Service Provider/query/")
+            .post("/v1/smartapi/978fe380a147a8641caf72320862697b/query/")
             .send(query)
             .set('Accept', 'application/json')
             .expect(200)
             .expect('Content-Type', /json/)
             .then(response => {
                 expect(response.body.message.knowledge_graph.nodes).toHaveProperty("CHEBI:32677")
-                expect(response.body.message.knowledge_graph.nodes["CHEBI:32677"].attributes[0].value).toEqual([
-                    "CHEBI:32677",
-                    "name:glutamine residue"
+                expect(response.body.message.knowledge_graph.nodes["CHEBI:32677"].attributes[0].value).toEqual(["CHEBI:32677"])
+            })
+    })
+
+    test("Query to non-Text Mining KPs should have id resolution turned on", async () => {
+        const query = JSON.parse(fs.readFileSync(path.join(example_foler, "serviceprovider/mygene.json")));
+        await request(app)
+            .post("/v1/smartapi/59dce17363dce279d389100834e43648/query")
+            .send(query)
+            .set('Accept', 'application/json')
+            .expect(200)
+            .expect('Content-Type', /json/)
+            .then(response => {
+                expect(response.body.message.knowledge_graph.nodes).toHaveProperty("GO:0000082")
+                expect(response.body.message.knowledge_graph.nodes["GO:0000082"].attributes[0].value).toEqual([
+                    "GO:0000082",
+                    "name:G1/S transition of mitotic cell cycle"
                 ])
             })
     })
@@ -97,7 +111,7 @@ describe("Testing /v1/team/{team_name}/query endpoints", () => {
         const apiResponse = await axios.get('https://biothings.ncats.io/text_mining_co_occurrence_kp/query?q=object.id:%22MONDO:0005252%22%20AND%20subject.type:%22ChemicalSubstance%22');
         const hits = apiResponse.data.total;
         await request(app)
-            .post("/v1/team/Text Mining Provider/query/")
+            .post("/v1/smartapi/5be0f321a829792e934545998b9c6afe/query/")
             .send(query)
             .set('Accept', 'application/json')
             .expect(200)
